+++ WebCharts.Services/Models/ChartTypes/ColumnChart.cs
@@ -367,12 +367,12 @@ namespace WebCharts.Services
                     {
                         // The formula for position is based on a distance
                         //from the grid line or nPoints position.
-                        xPosition = hAxis.GetPosition((double)index) - width * ((double)numOfSeries) / 2.0 + width / 2 + seriesIndx * width;
-                        xCenterVal = hAxis.GetPosition((double)index);
+                        xPosition = hAxis.GetPosition(index) - width * numOfSeries / 2.0 + width / 2 + seriesIndx * width;
+                        xCenterVal = hAxis.GetPosition(index);
                     }
                     else if (sameInterval)
                     {
-                        xPosition = hAxis.GetPosition(point.XValue) - width * ((double)numOfSeries) / 2.0 + width / 2 + seriesIndx * width;
+                        xPosition = hAxis.GetPosition(point.XValue) - width * numOfSeries / 2.0 + width / 2 + seriesIndx * width;
                         xCenterVal = hAxis.GetPosition(point.XValue);
                     }
                     else
